refactor(executor): Split batch.rs (1,300 lines) into logical modules
Split columnar/batch.rs into a module directory with focused files:
- types.rs (97 lines): Core ColumnarBatch/ColumnArray type definitions
- builder.rs (378 lines): Batch construction (new, from_rows, from_columns)
- operations.rs (273 lines): Accessors and manipulation methods
- arrow.rs (325 lines): Arrow RecordBatch conversion
- storage.rs (305 lines): Storage layer ColumnarTable conversion
- mod.rs (30 lines): Module documentation and re-exports

All modules are now under 400 lines for better maintainability.
No public API changes - all existing tests pass (178 columnar tests).

Closes #3107
