--- crates/vibesql-executor/src/select/columnar/batch.rs
@@ -1,1300 +1,0 @@
-//! Columnar batch structure for high-performance query execution
-//!
-//! This module implements true columnar storage with type-specialized column arrays
-//! optimized for SIMD operations and cache locality.
-//!
-//! ## Zero-Copy Design
-//!
-//! ColumnArray uses `Arc<Vec<T>>` for column data, enabling:
-//! - Zero-copy sharing with storage layer (via `from_storage_columnar`)
-//! - O(1) clone operations (reference count bump instead of data copy)
-//! - Cache-friendly data that can be shared across query executions
-//!
-//! Note: Index-based iteration is used in some places for performance (better vectorization).
-
-#![allow(clippy::needless_range_loop)]
-
-use std::sync::Arc;
-use crate::errors::ExecutorError;
-use vibesql_storage::Row;
-use vibesql_types::{DataType, SqlValue};
-
-use arrow::array::{Array, ArrayRef};
-use arrow::record_batch::RecordBatch;
-
-/// A columnar batch stores data in column-oriented format for efficient SIMD processing
-///
-/// Unlike row-oriented storage (Vec<Row>), columnar batches store each column
-/// in a contiguous array, enabling:
-/// - SIMD vectorization (process 4-8 values per instruction)
-/// - Better cache locality (columns accessed together are stored together)
-/// - Type-specialized code paths (no SqlValue enum matching)
-/// - Efficient NULL handling with separate bitmasks
-///
-/// # Example
-///
-/// ```rust,ignore
-/// // Convert rows to columnar batch
-/// let batch = ColumnarBatch::from_rows(&rows, &schema)?;
-///
-/// // Access columns with zero-copy
-/// if let ColumnArray::Int64(values, nulls) = &batch.columns[0] {
-///     // Process with SIMD operations
-///     let sum = simd_sum_i64(values);
-/// }
-/// ```
-#[derive(Debug, Clone)]
-pub struct ColumnarBatch {
-    /// Number of rows in this batch
-    row_count: usize,
-
-    /// Column arrays (one per column)
-    columns: Vec<ColumnArray>,
-
-    /// Optional column names for debugging
-    column_names: Option<Vec<String>>,
-}
-
-/// Type-specialized column storage
-///
-/// Each variant stores values in a native array for maximum SIMD efficiency.
-/// NULL values are tracked separately in a boolean bitmap.
-///
-/// ## Zero-Copy Design
-///
-/// All column data is wrapped in `Arc<Vec<T>>`, enabling:
-/// - Zero-copy sharing with storage layer
-/// - O(1) clone operations for query execution
-/// - Direct SIMD access via `as_slice()` on Arc contents
-#[derive(Debug, Clone)]
-pub enum ColumnArray {
-    /// 64-bit integers (INT, BIGINT)
-    Int64(Arc<Vec<i64>>, Option<Arc<Vec<bool>>>),
-
-    /// 32-bit integers (INT, SMALLINT)
-    Int32(Arc<Vec<i32>>, Option<Arc<Vec<bool>>>),
-
-    /// 64-bit floats (DOUBLE PRECISION, FLOAT)
-    Float64(Arc<Vec<f64>>, Option<Arc<Vec<bool>>>),
-
-    /// 32-bit floats (REAL)
-    Float32(Arc<Vec<f32>>, Option<Arc<Vec<bool>>>),
-
-    /// Variable-length strings (VARCHAR, TEXT)
-    String(Arc<Vec<String>>, Option<Arc<Vec<bool>>>),
-
-    /// Fixed-length strings (CHAR)
-    FixedString(Arc<Vec<String>>, Option<Arc<Vec<bool>>>),
-
-    /// Dates (stored as i32 days since epoch)
-    Date(Arc<Vec<i32>>, Option<Arc<Vec<bool>>>),
-
-    /// Timestamps (stored as i64 microseconds since epoch)
-    Timestamp(Arc<Vec<i64>>, Option<Arc<Vec<bool>>>),
-
-    /// Booleans (stored as bytes for SIMD compatibility)
-    Boolean(Arc<Vec<u8>>, Option<Arc<Vec<bool>>>),
-
-    /// Mixed-type column (fallback for complex types)
-    Mixed(Arc<Vec<SqlValue>>),
-}
-
-impl ColumnarBatch {
-    /// Create a new empty columnar batch
-    pub fn new(column_count: usize) -> Self {
-        Self {
-            row_count: 0,
-            columns: Vec::with_capacity(column_count),
-            column_names: None,
-        }
-    }
-
-    /// Create a columnar batch with specified capacity
-    pub fn with_capacity(_row_count: usize, column_count: usize) -> Self {
-        Self {
-            row_count: 0,
-            columns: Vec::with_capacity(column_count),
-            column_names: None,
-        }
-    }
-
-    /// Get the number of rows in this batch
-    pub fn row_count(&self) -> usize {
-        self.row_count
-    }
-
-    /// Get the number of columns in this batch
-    pub fn column_count(&self) -> usize {
-        self.columns.len()
-    }
-
-    /// Get a reference to a column array
-    pub fn column(&self, index: usize) -> Option<&ColumnArray> {
-        self.columns.get(index)
-    }
-
-    /// Get a mutable reference to a column array
-    pub fn column_mut(&mut self, index: usize) -> Option<&mut ColumnArray> {
-        self.columns.get_mut(index)
-    }
-
-    /// Add a column to the batch
-    pub fn add_column(&mut self, column: ColumnArray) -> Result<(), ExecutorError> {
-        // Verify column has correct length
-        let col_len = column.len();
-        if self.row_count > 0 && col_len != self.row_count {
-            return Err(ExecutorError::ColumnarLengthMismatch {
-                context: "add_column".to_string(),
-                expected: self.row_count,
-                actual: col_len,
-            });
-        }
-
-        if self.row_count == 0 {
-            self.row_count = col_len;
-        }
-
-        self.columns.push(column);
-        Ok(())
-    }
-
-    /// Set column names (for debugging)
-    pub fn set_column_names(&mut self, names: Vec<String>) {
-        self.column_names = Some(names);
-    }
-
-    /// Get column names
-    pub fn column_names(&self) -> Option<&[String]> {
-        self.column_names.as_deref()
-    }
-
-    /// Get column index by name
-    pub fn column_index_by_name(&self, name: &str) -> Option<usize> {
-        self.column_names.as_ref()?.iter().position(|n| n == name)
-    }
-
-    /// Convert from row-oriented storage to columnar batch
-    ///
-    /// This analyzes the first row to infer column types, then materializes
-    /// all values into type-specialized column arrays.
-    pub fn from_rows(rows: &[Row]) -> Result<Self, ExecutorError> {
-        if rows.is_empty() {
-            return Ok(Self::new(0));
-        }
-
-        let row_count = rows.len();
-        let column_count = rows[0].len();
-
-        // Infer column types from first row
-        let column_types = Self::infer_column_types(&rows[0]);
-
-        // Create column arrays
-        let mut columns = Vec::with_capacity(column_count);
-
-        for (col_idx, col_type) in column_types.iter().enumerate() {
-            let column = Self::extract_column(rows, col_idx, col_type)?;
-            columns.push(column);
-        }
-
-        Ok(Self {
-            row_count,
-            columns,
-            column_names: None,
-        })
-    }
-
-    /// Convert from Arrow RecordBatch to ColumnarBatch (zero-copy when possible)
-    ///
-    /// This provides integration with Arrow-based storage engines, enabling
-    /// zero-copy columnar query execution. Arrow's columnar format maps directly
-    /// to our ColumnarBatch structure.
-    ///
-    /// # Performance
-    ///
-    /// - **Zero-copy**: Numeric types (Int64, Float64) are converted with minimal overhead
-    /// - **< 1ms overhead**: Conversion time negligible compared to query execution
-    /// - **Memory efficient**: Reuses Arrow's allocated memory where possible
-    ///
-    /// # Arguments
-    ///
-    /// * `batch` - Arrow RecordBatch from storage layer
-    ///
-    /// # Returns
-    ///
-    /// A ColumnarBatch ready for SIMD-accelerated query execution
-    pub fn from_arrow_batch(batch: &RecordBatch) -> Result<Self, ExecutorError> {
-        let row_count = batch.num_rows();
-        let column_count = batch.num_columns();
-
-        let mut columns = Vec::with_capacity(column_count);
-        let mut column_names = Vec::with_capacity(column_count);
-
-        // Convert each Arrow column to our ColumnArray format
-        for (idx, field) in batch.schema().fields().iter().enumerate() {
-            column_names.push(field.name().clone());
-            let array = batch.column(idx);
-
-            let column = Self::convert_arrow_array(array, field.data_type())?;
-            columns.push(column);
-        }
-
-        Ok(Self {
-            row_count,
-            columns,
-            column_names: Some(column_names),
-        })
-    }
-
-    /// Convert a single Arrow array to ColumnArray
-    fn convert_arrow_array(
-        array: &ArrayRef,
-        data_type: &arrow::datatypes::DataType,
-    ) -> Result<ColumnArray, ExecutorError> {
-        use arrow::array::{
-            BooleanArray, Date32Array, Float32Array, Float64Array, Int32Array, Int64Array,
-            StringArray, TimestampMicrosecondArray,
-        };
-        use arrow::datatypes::DataType as ArrowDataType;
-
-        match data_type {
-            ArrowDataType::Int64 => {
-                let arr = array
-                    .as_any()
-                    .downcast_ref::<Int64Array>()
-                    .ok_or_else(|| ExecutorError::ArrowDowncastError {
-                        expected_type: "Int64Array".to_string(),
-                        context: "Arrow batch conversion".to_string(),
-                    })?;
-
-                let values: Vec<i64> = (0..arr.len()).map(|i| arr.value(i)).collect();
-                let nulls = if arr.null_count() > 0 {
-                    Some(Arc::new((0..arr.len()).map(|i| arr.is_null(i)).collect()))
-                } else {
-                    None
-                };
-
-                Ok(ColumnArray::Int64(Arc::new(values), nulls))
-            }
-
-            ArrowDataType::Int32 => {
-                let arr = array
-                    .as_any()
-                    .downcast_ref::<Int32Array>()
-                    .ok_or_else(|| ExecutorError::ArrowDowncastError {
-                        expected_type: "Int32Array".to_string(),
-                        context: "Arrow batch conversion".to_string(),
-                    })?;
-
-                let values: Vec<i32> = (0..arr.len()).map(|i| arr.value(i)).collect();
-                let nulls = if arr.null_count() > 0 {
-                    Some(Arc::new((0..arr.len()).map(|i| arr.is_null(i)).collect()))
-                } else {
-                    None
-                };
-
-                Ok(ColumnArray::Int32(Arc::new(values), nulls))
-            }
-
-            ArrowDataType::Float64 => {
-                let arr = array
-                    .as_any()
-                    .downcast_ref::<Float64Array>()
-                    .ok_or_else(|| ExecutorError::ArrowDowncastError {
-                        expected_type: "Float64Array".to_string(),
-                        context: "Arrow batch conversion".to_string(),
-                    })?;
-
-                let values: Vec<f64> = (0..arr.len()).map(|i| arr.value(i)).collect();
-                let nulls = if arr.null_count() > 0 {
-                    Some(Arc::new((0..arr.len()).map(|i| arr.is_null(i)).collect()))
-                } else {
-                    None
-                };
-
-                Ok(ColumnArray::Float64(Arc::new(values), nulls))
-            }
-
-            ArrowDataType::Float32 => {
-                let arr = array
-                    .as_any()
-                    .downcast_ref::<Float32Array>()
-                    .ok_or_else(|| ExecutorError::ArrowDowncastError {
-                        expected_type: "Float32Array".to_string(),
-                        context: "Arrow batch conversion".to_string(),
-                    })?;
-
-                let values: Vec<f32> = (0..arr.len()).map(|i| arr.value(i)).collect();
-                let nulls = if arr.null_count() > 0 {
-                    Some(Arc::new((0..arr.len()).map(|i| arr.is_null(i)).collect()))
-                } else {
-                    None
-                };
-
-                Ok(ColumnArray::Float32(Arc::new(values), nulls))
-            }
-
-            ArrowDataType::Utf8 => {
-                let arr = array
-                    .as_any()
-                    .downcast_ref::<StringArray>()
-                    .ok_or_else(|| ExecutorError::ArrowDowncastError {
-                        expected_type: "StringArray".to_string(),
-                        context: "Arrow batch conversion".to_string(),
-                    })?;
-
-                let values: Vec<String> = (0..arr.len()).map(|i| arr.value(i).to_string()).collect();
-                let nulls = if arr.null_count() > 0 {
-                    Some(Arc::new((0..arr.len()).map(|i| arr.is_null(i)).collect()))
-                } else {
-                    None
-                };
-
-                Ok(ColumnArray::String(Arc::new(values), nulls))
-            }
-
-            ArrowDataType::Boolean => {
-                let arr = array
-                    .as_any()
-                    .downcast_ref::<BooleanArray>()
-                    .ok_or_else(|| ExecutorError::ArrowDowncastError {
-                        expected_type: "BooleanArray".to_string(),
-                        context: "Arrow batch conversion".to_string(),
-                    })?;
-
-                let values: Vec<u8> = (0..arr.len())
-                    .map(|i| if arr.value(i) { 1 } else { 0 })
-                    .collect();
-                let nulls = if arr.null_count() > 0 {
-                    Some(Arc::new((0..arr.len()).map(|i| arr.is_null(i)).collect()))
-                } else {
-                    None
-                };
-
-                Ok(ColumnArray::Boolean(Arc::new(values), nulls))
-            }
-
-            ArrowDataType::Date32 => {
-                let arr = array
-                    .as_any()
-                    .downcast_ref::<Date32Array>()
-                    .ok_or_else(|| ExecutorError::ArrowDowncastError {
-                        expected_type: "Date32Array".to_string(),
-                        context: "Arrow batch conversion".to_string(),
-                    })?;
-
-                let values: Vec<i32> = (0..arr.len()).map(|i| arr.value(i)).collect();
-                let nulls = if arr.null_count() > 0 {
-                    Some(Arc::new((0..arr.len()).map(|i| arr.is_null(i)).collect()))
-                } else {
-                    None
-                };
-
-                Ok(ColumnArray::Date(Arc::new(values), nulls))
-            }
-
-            ArrowDataType::Timestamp(arrow::datatypes::TimeUnit::Microsecond, _) => {
-                let arr = array
-                    .as_any()
-                    .downcast_ref::<TimestampMicrosecondArray>()
-                    .ok_or_else(|| ExecutorError::ArrowDowncastError {
-                        expected_type: "TimestampMicrosecondArray".to_string(),
-                        context: "Arrow batch conversion".to_string(),
-                    })?;
-
-                let values: Vec<i64> = (0..arr.len()).map(|i| arr.value(i)).collect();
-                let nulls = if arr.null_count() > 0 {
-                    Some(Arc::new((0..arr.len()).map(|i| arr.is_null(i)).collect()))
-                } else {
-                    None
-                };
-
-                Ok(ColumnArray::Timestamp(Arc::new(values), nulls))
-            }
-
-            _ => Err(ExecutorError::UnsupportedArrayType {
-                operation: "Arrow batch conversion".to_string(),
-                array_type: format!("{:?}", data_type),
-            }),
-        }
-    }
-
-    /// Convert from storage layer ColumnarTable to executor ColumnarBatch
-    ///
-    /// This method provides **true zero-copy** conversion from the storage layer's
-    /// columnar format to the executor's columnar format. This is the key integration
-    /// point for native columnar table scans.
-    ///
-    /// # Performance
-    ///
-    /// - **O(1) for numeric/string columns**: Arc::clone is just a reference count bump
-    /// - **< 1 microsecond** for millions of rows (vs O(n) with data copy)
-    /// - Directly shares storage ColumnData with executor ColumnArray
-    /// - Critical path for TPC-H Q6 and other analytical queries
-    ///
-    /// # Zero-Copy Design
-    ///
-    /// Both `vibesql_storage::ColumnData` and executor `ColumnArray` use `Arc<Vec<T>>`
-    /// for column data. Calling `Arc::clone()` only increments a reference count,
-    /// avoiding any data copying:
-    ///
-    /// ```text
-    /// Storage: Arc<Vec<i64>> ─┬─> [1, 2, 3, 4, ...]  (shared memory)
-    ///                         │
-    /// Executor: Arc<Vec<i64>> ┘
-    /// ```
-    ///
-    /// # Arguments
-    ///
-    /// * `storage_columnar` - ColumnarTable from storage layer (vibesql-storage)
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(ColumnarBatch)` - Executor-ready columnar batch with shared Arc references
-    /// * `Err(ExecutorError)` - If type conversion fails
-    pub fn from_storage_columnar(
-        storage_columnar: &vibesql_storage::ColumnarTable,
-    ) -> Result<Self, ExecutorError> {
-        use vibesql_storage::ColumnData;
-
-        let column_names = storage_columnar.column_names().to_vec();
-        let row_count = storage_columnar.row_count();
-        let mut columns = Vec::with_capacity(column_names.len());
-
-        for col_name in column_names.iter() {
-            let storage_col = storage_columnar.get_column(col_name).ok_or_else(|| {
-                ExecutorError::ColumnarColumnNotFoundByName {
-                    column_name: col_name.clone(),
-                }
-            })?;
-
-            let column_array = match storage_col {
-                ColumnData::Int64 { values, nulls } => {
-                    // Zero-copy: Arc::clone is O(1) - just bumps reference count
-                    let null_bitmap = if nulls.iter().any(|&n| n) {
-                        Some(Arc::clone(nulls))
-                    } else {
-                        None
-                    };
-                    ColumnArray::Int64(Arc::clone(values), null_bitmap)
-                }
-                ColumnData::Float64 { values, nulls } => {
-                    // Zero-copy: Arc::clone is O(1)
-                    let null_bitmap = if nulls.iter().any(|&n| n) {
-                        Some(Arc::clone(nulls))
-                    } else {
-                        None
-                    };
-                    ColumnArray::Float64(Arc::clone(values), null_bitmap)
-                }
-                ColumnData::String { values, nulls } => {
-                    // Zero-copy: Arc::clone is O(1)
-                    let null_bitmap = if nulls.iter().any(|&n| n) {
-                        Some(Arc::clone(nulls))
-                    } else {
-                        None
-                    };
-                    ColumnArray::String(Arc::clone(values), null_bitmap)
-                }
-                ColumnData::Bool { values, nulls } => {
-                    // Convert bool to u8 for SIMD compatibility (requires iteration)
-                    let u8_values: Vec<u8> = values.iter().map(|&b| if b { 1 } else { 0 }).collect();
-                    let null_bitmap = if nulls.iter().any(|&n| n) {
-                        Some(Arc::clone(nulls))
-                    } else {
-                        None
-                    };
-                    ColumnArray::Boolean(Arc::new(u8_values), null_bitmap)
-                }
-                ColumnData::Date { values, nulls } => {
-                    // Convert Date to i32 (days since Unix epoch 1970-01-01)
-                    // Must use the same formula as simd_filter.rs:date_to_days_since_epoch
-                    // for predicate evaluation to work correctly
-                    let i32_values: Vec<i32> = values.iter().map(|d| {
-                        date_to_days_since_epoch(d)
-                    }).collect();
-                    let null_bitmap = if nulls.iter().any(|&n| n) {
-                        Some(Arc::clone(nulls))
-                    } else {
-                        None
-                    };
-                    ColumnArray::Date(Arc::new(i32_values), null_bitmap)
-                }
-                ColumnData::Timestamp { values, nulls } => {
-                    // Convert Timestamp to Mixed (fallback - no direct i64 conversion)
-                    let sql_values: Vec<SqlValue> = values
-                        .iter()
-                        .zip(nulls.iter())
-                        .map(|(t, &is_null)| {
-                            if is_null {
-                                SqlValue::Null
-                            } else {
-                                SqlValue::Timestamp(*t)
-                            }
-                        })
-                        .collect();
-                    ColumnArray::Mixed(Arc::new(sql_values))
-                }
-                ColumnData::Time { values, nulls } => {
-                    // Convert Time to Mixed (fallback - Time doesn't have direct i64 conversion)
-                    let sql_values: Vec<SqlValue> = values
-                        .iter()
-                        .zip(nulls.iter())
-                        .map(|(t, &is_null)| {
-                            if is_null {
-                                SqlValue::Null
-                            } else {
-                                SqlValue::Time(*t)
-                            }
-                        })
-                        .collect();
-                    ColumnArray::Mixed(Arc::new(sql_values))
-                }
-                ColumnData::Interval { values, nulls } => {
-                    // Convert Interval to Mixed (fallback)
-                    let sql_values: Vec<SqlValue> = values
-                        .iter()
-                        .zip(nulls.iter())
-                        .map(|(i, &is_null)| {
-                            if is_null {
-                                SqlValue::Null
-                            } else {
-                                SqlValue::Interval(i.clone())
-                            }
-                        })
-                        .collect();
-                    ColumnArray::Mixed(Arc::new(sql_values))
-                }
-            };
-
-            columns.push(column_array);
-        }
-
-        Ok(Self {
-            row_count,
-            columns,
-            column_names: Some(column_names),
-        })
-    }
-
-    /// Convert columnar batch back to row-oriented storage
-    pub fn to_rows(&self) -> Result<Vec<Row>, ExecutorError> {
-        let mut rows = Vec::with_capacity(self.row_count);
-
-        for row_idx in 0..self.row_count {
-            let mut values = Vec::with_capacity(self.columns.len());
-
-            for column in &self.columns {
-                let value = column.get_value(row_idx)?;
-                values.push(value);
-            }
-
-            rows.push(Row::new(values));
-        }
-
-        Ok(rows)
-    }
-
-    /// Get a value at a specific (row, column) position
-    pub fn get_value(&self, row_idx: usize, col_idx: usize) -> Result<SqlValue, ExecutorError> {
-        let column = self.column(col_idx)
-            .ok_or(ExecutorError::ColumnarColumnNotFound {
-                column_index: col_idx,
-                batch_columns: self.columns.len(),
-            })?;
-        column.get_value(row_idx)
-    }
-
-    /// Create an empty batch with the specified number of columns
-    pub fn empty(column_count: usize) -> Result<Self, ExecutorError> {
-        Ok(Self {
-            row_count: 0,
-            columns: vec![ColumnArray::Mixed(Arc::new(vec![])); column_count],
-            column_names: None,
-        })
-    }
-
-    /// Create a batch from a list of columns
-    pub fn from_columns(
-        columns: Vec<ColumnArray>,
-        column_names: Option<Vec<String>>,
-    ) -> Result<Self, ExecutorError> {
-        if columns.is_empty() {
-            return Ok(Self {
-                row_count: 0,
-                columns,
-                column_names,
-            });
-        }
-
-        // Verify all columns have the same length
-        let row_count = columns[0].len();
-        for (idx, column) in columns.iter().enumerate() {
-            if column.len() != row_count {
-                return Err(ExecutorError::ColumnarLengthMismatch {
-                    context: format!("from_columns (column {})", idx),
-                    expected: row_count,
-                    actual: column.len(),
-                });
-            }
-        }
-
-        Ok(Self {
-            row_count,
-            columns,
-            column_names,
-        })
-    }
-
-    /// Extract a single column from rows into a typed array
-    fn extract_column(
-        rows: &[Row],
-        col_idx: usize,
-        col_type: &ColumnType,
-    ) -> Result<ColumnArray, ExecutorError> {
-        match col_type {
-            ColumnType::Int64 => {
-                let mut values = Vec::with_capacity(rows.len());
-                let mut nulls = Vec::with_capacity(rows.len());
-                let mut has_nulls = false;
-
-                for row in rows {
-                    match row.get(col_idx) {
-                        Some(SqlValue::Integer(v)) => {
-                            values.push(*v);
-                            nulls.push(false);
-                        }
-                        Some(SqlValue::Null) => {
-                            values.push(0); // placeholder
-                            nulls.push(true);
-                            has_nulls = true;
-                        }
-                        Some(other) => {
-                            return Err(ExecutorError::ColumnarTypeMismatch {
-                                operation: "extract_column".to_string(),
-                                left_type: "Integer".to_string(),
-                                right_type: Some(format!("{:?}", other)),
-                            });
-                        }
-                        None => {
-                            values.push(0);
-                            nulls.push(true);
-                            has_nulls = true;
-                        }
-                    }
-                }
-
-                Ok(ColumnArray::Int64(Arc::new(values), if has_nulls { Some(Arc::new(nulls)) } else { None }))
-            }
-
-            ColumnType::Float64 => {
-                let mut values = Vec::with_capacity(rows.len());
-                let mut nulls = Vec::with_capacity(rows.len());
-                let mut has_nulls = false;
-
-                for row in rows {
-                    match row.get(col_idx) {
-                        Some(SqlValue::Double(v)) => {
-                            values.push(*v);
-                            nulls.push(false);
-                        }
-                        Some(SqlValue::Null) => {
-                            values.push(0.0); // placeholder
-                            nulls.push(true);
-                            has_nulls = true;
-                        }
-                        Some(other) => {
-                            return Err(ExecutorError::ColumnarTypeMismatch {
-                                operation: "extract_column".to_string(),
-                                left_type: "Double".to_string(),
-                                right_type: Some(format!("{:?}", other)),
-                            });
-                        }
-                        None => {
-                            values.push(0.0);
-                            nulls.push(true);
-                            has_nulls = true;
-                        }
-                    }
-                }
-
-                Ok(ColumnArray::Float64(Arc::new(values), if has_nulls { Some(Arc::new(nulls)) } else { None }))
-            }
-
-            ColumnType::String => {
-                let mut values = Vec::with_capacity(rows.len());
-                let mut nulls = Vec::with_capacity(rows.len());
-                let mut has_nulls = false;
-
-                for row in rows {
-                    match row.get(col_idx) {
-                        Some(SqlValue::Varchar(v)) => {
-                            values.push(v.clone());
-                            nulls.push(false);
-                        }
-                        Some(SqlValue::Null) => {
-                            values.push(String::new()); // placeholder
-                            nulls.push(true);
-                            has_nulls = true;
-                        }
-                        Some(other) => {
-                            return Err(ExecutorError::ColumnarTypeMismatch {
-                                operation: "extract_column".to_string(),
-                                left_type: "Varchar".to_string(),
-                                right_type: Some(format!("{:?}", other)),
-                            });
-                        }
-                        None => {
-                            values.push(String::new());
-                            nulls.push(true);
-                            has_nulls = true;
-                        }
-                    }
-                }
-
-                Ok(ColumnArray::String(Arc::new(values), if has_nulls { Some(Arc::new(nulls)) } else { None }))
-            }
-
-            ColumnType::Date | ColumnType::Mixed => {
-                // Store dates and mixed types as Mixed (fallback for non-SIMD types)
-                let mut values = Vec::with_capacity(rows.len());
-
-                for row in rows {
-                    let value = row.get(col_idx).cloned().unwrap_or(SqlValue::Null);
-                    values.push(value);
-                }
-
-                Ok(ColumnArray::Mixed(Arc::new(values)))
-            }
-
-            ColumnType::Boolean => {
-                let mut values = Vec::with_capacity(rows.len());
-                let mut nulls = Vec::with_capacity(rows.len());
-                let mut has_nulls = false;
-
-                for row in rows {
-                    match row.get(col_idx) {
-                        Some(SqlValue::Boolean(b)) => {
-                            values.push(if *b { 1 } else { 0 });
-                            nulls.push(false);
-                        }
-                        Some(SqlValue::Null) => {
-                            values.push(0); // placeholder
-                            nulls.push(true);
-                            has_nulls = true;
-                        }
-                        Some(other) => {
-                            return Err(ExecutorError::ColumnarTypeMismatch {
-                                operation: "extract_column".to_string(),
-                                left_type: "Boolean".to_string(),
-                                right_type: Some(format!("{:?}", other)),
-                            });
-                        }
-                        None => {
-                            values.push(0);
-                            nulls.push(true);
-                            has_nulls = true;
-                        }
-                    }
-                }
-
-                Ok(ColumnArray::Boolean(Arc::new(values), if has_nulls { Some(Arc::new(nulls)) } else { None }))
-            }
-        }
-    }
-
-    /// Infer column types from the first row
-    fn infer_column_types(first_row: &Row) -> Vec<ColumnType> {
-        let mut types = Vec::with_capacity(first_row.len());
-
-        for i in 0..first_row.len() {
-            let col_type = match first_row.get(i) {
-                Some(SqlValue::Integer(_)) => ColumnType::Int64,
-                Some(SqlValue::Double(_)) => ColumnType::Float64,
-                Some(SqlValue::Varchar(_)) => ColumnType::String,
-                Some(SqlValue::Date(_)) => ColumnType::Date,
-                Some(SqlValue::Boolean(_)) => ColumnType::Boolean,
-                _ => ColumnType::Mixed,
-            };
-            types.push(col_type);
-        }
-
-        types
-    }
-}
-
-/// Internal column type inference
-#[derive(Debug, Clone, Copy)]
-enum ColumnType {
-    Int64,
-    Float64,
-    String,
-    Date,
-    Boolean,
-    Mixed,
-}
-
-impl ColumnArray {
-    /// Get the number of values in this column
-    pub fn len(&self) -> usize {
-        match self {
-            Self::Int64(v, _) => v.len(),
-            Self::Int32(v, _) => v.len(),
-            Self::Float64(v, _) => v.len(),
-            Self::Float32(v, _) => v.len(),
-            Self::String(v, _) => v.len(),
-            Self::FixedString(v, _) => v.len(),
-            Self::Date(v, _) => v.len(),
-            Self::Timestamp(v, _) => v.len(),
-            Self::Boolean(v, _) => v.len(),
-            Self::Mixed(v) => v.len(),
-        }
-    }
-
-    /// Check if column is empty
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// Get a value at the specified index as SqlValue
-    pub fn get_value(&self, index: usize) -> Result<SqlValue, ExecutorError> {
-        match self {
-            Self::Int64(values, nulls) => {
-                if let Some(null_mask) = nulls {
-                    if null_mask.get(index).copied().unwrap_or(false) {
-                        return Ok(SqlValue::Null);
-                    }
-                }
-                values.get(index)
-                    .map(|v| SqlValue::Integer(*v))
-                    .ok_or(ExecutorError::ColumnIndexOutOfBounds { index })
-            }
-
-            Self::Float64(values, nulls) => {
-                if let Some(null_mask) = nulls {
-                    if null_mask.get(index).copied().unwrap_or(false) {
-                        return Ok(SqlValue::Null);
-                    }
-                }
-                values.get(index)
-                    .map(|v| SqlValue::Double(*v))
-                    .ok_or(ExecutorError::ColumnIndexOutOfBounds { index })
-            }
-
-            Self::String(values, nulls) => {
-                if let Some(null_mask) = nulls {
-                    if null_mask.get(index).copied().unwrap_or(false) {
-                        return Ok(SqlValue::Null);
-                    }
-                }
-                values.get(index)
-                    .map(|v| SqlValue::Varchar(v.clone()))
-                    .ok_or(ExecutorError::ColumnIndexOutOfBounds { index })
-            }
-
-            Self::Boolean(values, nulls) => {
-                if let Some(null_mask) = nulls {
-                    if null_mask.get(index).copied().unwrap_or(false) {
-                        return Ok(SqlValue::Null);
-                    }
-                }
-                values.get(index)
-                    .map(|v| SqlValue::Boolean(*v != 0))
-                    .ok_or(ExecutorError::ColumnIndexOutOfBounds { index })
-            }
-
-            Self::Mixed(values) => {
-                values.get(index)
-                    .cloned()
-                    .ok_or(ExecutorError::ColumnIndexOutOfBounds { index })
-            }
-
-            _ => Err(ExecutorError::UnsupportedArrayType {
-                operation: "get_value".to_string(),
-                array_type: format!("{:?}", std::mem::discriminant(self)),
-            }),
-        }
-    }
-
-    /// Get the data type of this column
-    pub fn data_type(&self) -> DataType {
-        match self {
-            Self::Int64(_, _) => DataType::Integer,
-            Self::Int32(_, _) => DataType::Integer,
-            Self::Float64(_, _) => DataType::DoublePrecision,
-            Self::Float32(_, _) => DataType::Real,
-            Self::String(_, _) => DataType::Varchar { max_length: None },
-            Self::FixedString(_, _) => DataType::Character { length: 255 },
-            Self::Date(_, _) => DataType::Date,
-            Self::Timestamp(_, _) => DataType::Timestamp { with_timezone: false },
-            Self::Boolean(_, _) => DataType::Boolean,
-            Self::Mixed(_) => DataType::Varchar { max_length: None }, // fallback
-        }
-    }
-
-    /// Get raw i64 slice (for SIMD operations)
-    pub fn as_i64(&self) -> Option<(&[i64], Option<&[bool]>)> {
-        match self {
-            Self::Int64(values, nulls) => Some((values.as_slice(), nulls.as_ref().map(|n| n.as_slice()))),
-            _ => None,
-        }
-    }
-
-    /// Get raw f64 slice (for SIMD operations)
-    pub fn as_f64(&self) -> Option<(&[f64], Option<&[bool]>)> {
-        match self {
-            Self::Float64(values, nulls) => Some((values.as_slice(), nulls.as_ref().map(|n| n.as_slice()))),
-            _ => None,
-        }
-    }
-}
-
-/// Convert Date to days since Unix epoch (1970-01-01)
-///
-/// This function MUST be kept in sync with simd_filter.rs::date_to_days_since_epoch()
-/// to ensure predicates compare dates correctly.
-fn date_to_days_since_epoch(date: &vibesql_types::Date) -> i32 {
-    // Accurate days since Unix epoch calculation with leap year handling
-    let year_days = (date.year - 1970) * 365;
-    let leap_years = ((date.year - 1969) / 4) - ((date.year - 1901) / 100) + ((date.year - 1601) / 400);
-    let month_days: i32 = match date.month {
-        1 => 0,
-        2 => 31,
-        3 => 59,
-        4 => 90,
-        5 => 120,
-        6 => 151,
-        7 => 181,
-        8 => 212,
-        9 => 243,
-        10 => 273,
-        11 => 304,
-        12 => 334,
-        _ => 0,
-    };
-
-    // Add leap day if after February in a leap year
-    let is_leap = date.year % 4 == 0 && (date.year % 100 != 0 || date.year % 400 == 0);
-    let leap_adjustment = if is_leap && date.month > 2 { 1 } else { 0 };
-
-    year_days + leap_years + month_days + date.day as i32 - 1 + leap_adjustment
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_columnar_batch_creation() {
-        let rows = vec![
-            Row::new(vec![
-                SqlValue::Integer(1),
-                SqlValue::Double(10.5),
-                SqlValue::Varchar("Alice".to_string()),
-            ]),
-            Row::new(vec![
-                SqlValue::Integer(2),
-                SqlValue::Double(20.5),
-                SqlValue::Varchar("Bob".to_string()),
-            ]),
-            Row::new(vec![
-                SqlValue::Integer(3),
-                SqlValue::Double(30.5),
-                SqlValue::Varchar("Charlie".to_string()),
-            ]),
-        ];
-
-        let batch = ColumnarBatch::from_rows(&rows).unwrap();
-
-        assert_eq!(batch.row_count(), 3);
-        assert_eq!(batch.column_count(), 3);
-
-        // Check column 0 (integers)
-        let col0 = batch.column(0).unwrap();
-        if let ColumnArray::Int64(values, nulls) = col0 {
-            assert_eq!(values.as_slice(), &[1, 2, 3]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected Int64 column");
-        }
-
-        // Check column 1 (doubles)
-        let col1 = batch.column(1).unwrap();
-        if let ColumnArray::Float64(values, nulls) = col1 {
-            assert_eq!(values.as_slice(), &[10.5, 20.5, 30.5]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected Float64 column");
-        }
-
-        // Check column 2 (strings)
-        let col2 = batch.column(2).unwrap();
-        if let ColumnArray::String(values, nulls) = col2 {
-            assert_eq!(values.as_slice(), &["Alice", "Bob", "Charlie"]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected String column");
-        }
-    }
-
-    #[test]
-    fn test_columnar_batch_with_nulls() {
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Double(10.0)]),
-            Row::new(vec![SqlValue::Null, SqlValue::Double(20.0)]),
-            Row::new(vec![SqlValue::Integer(3), SqlValue::Null]),
-        ];
-
-        let batch = ColumnarBatch::from_rows(&rows).unwrap();
-
-        // Check column 0 (with NULL)
-        let col0 = batch.column(0).unwrap();
-        if let ColumnArray::Int64(values, Some(nulls)) = col0 {
-            assert_eq!(values.len(), 3);
-            assert_eq!(nulls.as_slice(), &[false, true, false]);
-        } else {
-            panic!("Expected Int64 column with nulls");
-        }
-
-        // Check column 1 (with NULL)
-        let col1 = batch.column(1).unwrap();
-        if let ColumnArray::Float64(values, Some(nulls)) = col1 {
-            assert_eq!(values.len(), 3);
-            assert_eq!(nulls.as_slice(), &[false, false, true]);
-        } else {
-            panic!("Expected Float64 column with nulls");
-        }
-    }
-
-    #[test]
-    fn test_batch_to_rows_roundtrip() {
-        let original_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Double(10.5)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Double(20.5)]),
-        ];
-
-        let batch = ColumnarBatch::from_rows(&original_rows).unwrap();
-        let converted_rows = batch.to_rows().unwrap();
-
-        assert_eq!(converted_rows.len(), original_rows.len());
-        for (original, converted) in original_rows.iter().zip(converted_rows.iter()) {
-            assert_eq!(original.len(), converted.len());
-            for i in 0..original.len() {
-                assert_eq!(original.get(i), converted.get(i));
-            }
-        }
-    }
-
-    #[test]
-    fn test_simd_column_access() {
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Double(10.5)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Double(20.5)]),
-            Row::new(vec![SqlValue::Integer(3), SqlValue::Double(30.5)]),
-        ];
-
-        let batch = ColumnarBatch::from_rows(&rows).unwrap();
-
-        // Access i64 column for SIMD
-        let col0 = batch.column(0).unwrap();
-        if let Some((values, nulls)) = col0.as_i64() {
-            assert_eq!(values, &[1, 2, 3]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected i64 slice");
-        }
-
-        // Access f64 column for SIMD
-        let col1 = batch.column(1).unwrap();
-        if let Some((values, nulls)) = col1.as_f64() {
-            assert_eq!(values, &[10.5, 20.5, 30.5]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected f64 slice");
-        }
-    }
-
-    #[test]
-    fn test_arrow_integration() {
-        use arrow::array::{Float64Array, Int64Array};
-        use arrow::datatypes::{DataType as ArrowDataType, Field, Schema};
-        use arrow::record_batch::RecordBatch;
-        use std::sync::Arc;
-
-        // Create Arrow RecordBatch
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("id", ArrowDataType::Int64, false),
-            Field::new("value", ArrowDataType::Float64, false),
-        ]));
-
-        let id_array = Arc::new(Int64Array::from(vec![1, 2, 3]));
-        let value_array = Arc::new(Float64Array::from(vec![10.5, 20.5, 30.5]));
-
-        let arrow_batch =
-            RecordBatch::try_new(schema.clone(), vec![id_array, value_array]).unwrap();
-
-        // Convert to ColumnarBatch
-        let columnar_batch = ColumnarBatch::from_arrow_batch(&arrow_batch).unwrap();
-
-        // Verify structure
-        assert_eq!(columnar_batch.row_count(), 3);
-        assert_eq!(columnar_batch.column_count(), 2);
-
-        // Verify column names
-        let names = columnar_batch.column_names().unwrap();
-        assert_eq!(names, &["id", "value"]);
-
-        // Verify Int64 column
-        let col0 = columnar_batch.column(0).unwrap();
-        if let Some((values, nulls)) = col0.as_i64() {
-            assert_eq!(values, &[1, 2, 3]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected i64 column");
-        }
-
-        // Verify Float64 column
-        let col1 = columnar_batch.column(1).unwrap();
-        if let Some((values, nulls)) = col1.as_f64() {
-            assert_eq!(values, &[10.5, 20.5, 30.5]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected f64 column");
-        }
-    }
-
-    #[test]
-    fn test_arrow_integration_with_nulls() {
-        use arrow::array::{Float64Array, Int64Array};
-        use arrow::datatypes::{DataType as ArrowDataType, Field, Schema};
-        use arrow::record_batch::RecordBatch;
-        use std::sync::Arc;
-
-        // Create Arrow RecordBatch with NULLs
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("id", ArrowDataType::Int64, true), // nullable
-            Field::new("value", ArrowDataType::Float64, true),
-        ]));
-
-        let id_array = Arc::new(Int64Array::from(vec![Some(1), None, Some(3)]));
-        let value_array = Arc::new(Float64Array::from(vec![Some(10.5), Some(20.5), None]));
-
-        let arrow_batch =
-            RecordBatch::try_new(schema.clone(), vec![id_array, value_array]).unwrap();
-
-        // Convert to ColumnarBatch
-        let columnar_batch = ColumnarBatch::from_arrow_batch(&arrow_batch).unwrap();
-
-        // Verify Int64 column with NULL
-        let col0 = columnar_batch.column(0).unwrap();
-        if let Some((values, Some(nulls))) = col0.as_i64() {
-            assert_eq!(values.len(), 3);
-            assert_eq!(nulls, &[false, true, false]);
-        } else {
-            panic!("Expected i64 column with nulls");
-        }
-
-        // Verify Float64 column with NULL
-        let col1 = columnar_batch.column(1).unwrap();
-        if let Some((values, Some(nulls))) = col1.as_f64() {
-            assert_eq!(values.len(), 3);
-            assert_eq!(nulls, &[false, false, true]);
-        } else {
-            panic!("Expected f64 column with nulls");
-        }
-    }
-
-    #[test]
-    fn test_from_storage_columnar() {
-        // Create storage-layer columnar table
-        let rows = vec![
-            Row::new(vec![
-                SqlValue::Integer(1),
-                SqlValue::Double(10.5),
-                SqlValue::Varchar("Alice".to_string()),
-            ]),
-            Row::new(vec![
-                SqlValue::Integer(2),
-                SqlValue::Double(20.5),
-                SqlValue::Varchar("Bob".to_string()),
-            ]),
-            Row::new(vec![
-                SqlValue::Integer(3),
-                SqlValue::Double(30.5),
-                SqlValue::Varchar("Charlie".to_string()),
-            ]),
-        ];
-        let column_names = vec!["id".to_string(), "value".to_string(), "name".to_string()];
-        let storage_columnar = vibesql_storage::ColumnarTable::from_rows(&rows, &column_names).unwrap();
-
-        // Convert to executor ColumnarBatch
-        let batch = ColumnarBatch::from_storage_columnar(&storage_columnar).unwrap();
-
-        // Verify structure
-        assert_eq!(batch.row_count(), 3);
-        assert_eq!(batch.column_count(), 3);
-
-        // Verify column names
-        let names = batch.column_names().unwrap();
-        assert_eq!(names, &["id", "value", "name"]);
-
-        // Verify Int64 column
-        let col0 = batch.column(0).unwrap();
-        if let Some((values, nulls)) = col0.as_i64() {
-            assert_eq!(values, &[1, 2, 3]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected i64 column");
-        }
-
-        // Verify Float64 column
-        let col1 = batch.column(1).unwrap();
-        if let Some((values, nulls)) = col1.as_f64() {
-            assert_eq!(values, &[10.5, 20.5, 30.5]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected f64 column");
-        }
-
-        // Verify String column
-        let col2 = batch.column(2).unwrap();
-        if let ColumnArray::String(values, nulls) = col2 {
-            assert_eq!(values.as_slice(), &["Alice", "Bob", "Charlie"]);
-            assert!(nulls.is_none());
-        } else {
-            panic!("Expected String column");
-        }
-    }
-
-    #[test]
-    fn test_from_storage_columnar_with_nulls() {
-        // Create storage-layer columnar table with NULLs
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Double(10.0)]),
-            Row::new(vec![SqlValue::Null, SqlValue::Double(20.0)]),
-            Row::new(vec![SqlValue::Integer(3), SqlValue::Null]),
-        ];
-        let column_names = vec!["id".to_string(), "value".to_string()];
-        let storage_columnar = vibesql_storage::ColumnarTable::from_rows(&rows, &column_names).unwrap();
-
-        // Convert to executor ColumnarBatch
-        let batch = ColumnarBatch::from_storage_columnar(&storage_columnar).unwrap();
-
-        // Verify Int64 column with NULL
-        let col0 = batch.column(0).unwrap();
-        if let Some((values, Some(nulls))) = col0.as_i64() {
-            assert_eq!(values.len(), 3);
-            assert_eq!(nulls, &[false, true, false]);
-        } else {
-            panic!("Expected i64 column with nulls");
-        }
-
-        // Verify Float64 column with NULL
-        let col1 = batch.column(1).unwrap();
-        if let Some((values, Some(nulls))) = col1.as_f64() {
-            assert_eq!(values.len(), 3);
-            assert_eq!(nulls, &[false, false, true]);
-        } else {
-            panic!("Expected f64 column with nulls");
-        }
-    }
-}
